Typ Verrechnungskonto Property (#972)

+++ src/de/jost_net/JVerein/gui/control/EinstellungControl.java
@@ -2452,7 +2452,7 @@ public class EinstellungControl extends AbstractControl
       if (verrechnungskonto.getValue() != null)
       {
         Einstellungen.setEinstellung(Property.VERRECHNUNGSKONTOID,
-            (Long.parseLong(
+            (Integer.parseInt(
                 (String) ((Konto) verrechnungskonto.getValue()).getID())));
       }
       DBTransaction.commit();
